+++ tile.c
@@ -18,7 +18,7 @@
  * +--------+
  *
  */
-void monocle(const monitor_t *mon, unsigned int wins)
+void monocle(const monitor_t *mon, unsigned short wins)
 {
     for (client_t *c = cfg.clients; c && wins; c = c->next)
         if (ON_MONITOR(mon, c) && IS_VISIBLE(c) && IS_TILED(c)) {
@@ -37,11 +37,11 @@ void monocle(const monitor_t *mon, unsigned int wins)
  * +--------+
  *
  */
-void grid(const monitor_t *mon, unsigned int wins)
+void grid(const monitor_t *mon, unsigned short wins)
 {
-    const unsigned int n = wins;
-    unsigned int cols = 0, rows = 0;
-    unsigned int ncol = 0, nrow = 0;
+    const unsigned short n = wins;
+    unsigned short cols = 0, rows = 0;
+    unsigned short ncol = 0, nrow = 0;
 
     /* calculate number of columns and rows - emulate square root */
     for (cols = 0; cols <= n/2; cols++)
@@ -51,7 +51,7 @@ void grid(const monitor_t *mon, unsigned int wins)
         cols = 2;
     rows = n/cols;
 
-    unsigned int i = 0;
+    unsigned short i = 0;
 
     const uint16_t client_width  = mon->geom.width / (cols ? cols : 1);
     const uint16_t client_height = mon->geom.height;
@@ -83,7 +83,7 @@ void grid(const monitor_t *mon, unsigned int wins)
  * +--------+
  *
  */
-void vstack(const monitor_t *mon, unsigned int wins)
+void vstack(const monitor_t *mon, unsigned short wins)
 {
     client_t *c = cfg.clients;
 
@@ -91,7 +91,7 @@ void vstack(const monitor_t *mon, unsigned int wins)
     const uint16_t m_area = mon->geom.width * M_AREA_FACT + mon->m_area;
     const uint16_t m_wins = (wins <= mon->m_wins) ? wins - 1 : mon->m_wins;
 
-    for (int i = 0, m_h = mon->geom.height / m_wins; c && i < m_wins; c = c->next)
+    for (unsigned short i = 0, m_h = mon->geom.height / m_wins; c && i < m_wins; c = c->next)
         if (ON_MONITOR(mon, c) && IS_VISIBLE(c) && IS_TILED(c))
             client_set_geom(c, mon->geom.x, mon->geom.y + i++ * m_h, m_area, m_h);
 
@@ -120,7 +120,7 @@ void vstack(const monitor_t *mon, unsigned int wins)
  * +--------+
  *
  */
-void hstack(const monitor_t *mon, unsigned int wins)
+void hstack(const monitor_t *mon, unsigned short wins)
 {
     client_t *c = cfg.clients;
 
@@ -128,7 +128,7 @@ void hstack(const monitor_t *mon, unsigned int wins)
     const uint16_t m_area = mon->geom.height * M_AREA_FACT + mon->m_area;
     const uint16_t m_wins = (wins <= mon->m_wins) ? wins - 1 : mon->m_wins;
 
-    for (int i = 0, m_w = mon->geom.width / m_wins; c && i < m_wins; c = c->next)
+    for (unsigned short i = 0, m_w = mon->geom.width / m_wins; c && i < m_wins; c = c->next)
         if (ON_MONITOR(mon, c) && IS_VISIBLE(c) && IS_TILED(c))
             client_set_geom(c, mon->geom.x + i++ * m_w, mon->geom.y, m_w, m_area);
 
@@ -152,7 +152,7 @@ void hstack(const monitor_t *mon, unsigned int wins)
  */
 void tile(const monitor_t *mon, layout_t layout)
 {
-    unsigned int num_windows = 0;
+    unsigned short num_windows = 0;
     for (client_t *c = cfg.clients; c; c = c->next)
         if (ON_MONITOR(mon, c) && IS_VISIBLE(c) && IS_TILED(c))
             ++num_windows;
